refactor: 클래스 도메인 분류
